requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
Markdown==3.2.1
prettyconf==2.1.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
unicodecsv==0.14.1
urllib3==1.25.9
whitenoise==5.0.1
